No need to check parent path components when creating a directory.
Given S3 semantics, there's actually no problem with having a/b/c be a prefix even if
a/b or a is already a file. So there's no need to check for it - it wastes API calls
and can lead to problems with access control if the caller only has permissions
starting at some prefix.