Add support for storing link costs by invoking rack awareness script
The NetworkTopology can now store costs between a pair of racks. The
DatanodeManager invokes the rack awareness script with the "--cost"
switch, when a new datanode is registered. This allows it to store the
costs between each pair of racks.

This basic cost matrix can be used later to make GDA aware decisions