MAPREDUCE-6931. Remove TestDFSIO "Total Throughput" calculation.
Previously it failed to convert ms to seconds and thus reports aggregate
throughput as 1/1000x actual numbers. By recommendation from reviewers,
such a calculation is not in-scope for TestDFSIO anyways, so remove it.
Also, make all the bytes-to-mb and milliseconds-to-seconds conversions
consistent in the reporting messages to help avoid this type of error
in the future.