YARN-8246 winutils - fix failure to retrieve disk and network perf counters on localized windows installlation
PdhAddCounter expects performance counter path to be in the same language than the windows installation.
With current code, the calls to PdhAddCounter will fail with error 0xc0000bb8 (PDH_CSTATUS_NO_OBJECT)
on non-english windows installation.

The solution is to use PdhAddEnglishCounter function instead
https://docs.microsoft.com/en-us/windows/desktop/api/pdh/nf-pdh-pdhaddenglishcounterw

"This function provides a language-neutral way to add performance counters to the query"