HADOOP-16190. S3A copyFile operation to include source versionID or etag in the copy request
This patch adds the constraints on the request, and maps a 412 response to a RemoteFileChangedException.

No obvious test for this. The way to do it would be to get an invalid etag/version in to the request and see what happens, which would complicate the copy API a bit -but is something we will need for etag/version tracking in s3guard anyway....

Change-Id: I4b229336ba2d57018bd8b66888b807074419598e