HADOOP-16090 - reduced parental existence checks
This version of the patch attempts to reduce the #of read requests by only looking at the parent dir.

It doesn't work, as it doesn't  handle the use case: mkdir 2+ directories under an existing empty directory marker.
ITestS3AFileOperationCost.testFakeDirectoryDeletion fails for this reason.

mkdirs() will have to do a full walk & delete, though it could actually be done during the mkdir operation, which walks up anyway.

for file creation, the probe-and-delete for the parent should be enough. In day to day use, most applications make a point of creating a directory before opening a file, even though we never actually check this fact (or indeed, that you don't create a file under a file)

Change-Id: I404427cd9f310990450f3b2fc155bedb45c3de33