Fix S3AFileStatus group handling
ITestS3AConfiguration.testUsernameFromUGI was failing, expecting the
user to be copied into the group.

Strict copying of user into group causes
TestLocalMetadataStore.testPutNew() to fail since it expects the group
to be preserved from the original FileStatus.

This change copies user into group when group is null/empty. With this
change, all existing tests pass.