HADOOP-16118. S3Guard to support on-demand DDB tables.
This is the first step for on-demand operations: things recognize when they are using on-demand tables,
as do the tests.

Branch-3.2 patch; test run failing for unrelated changes to the AbstractS3GuardToolTestBase test suite,
which highlight that we have enough S3Guard changes in trunk it's time to backport one by one.

Contributed by Steve Loughran.

Change-Id: I44cf87a2bd253b78c2d0414e4be20e920027c980