HADOOP-15183: rename operation failure handling works in the test cases
* Add explicit notion of a StoreOperation, similar to Gang-of-four Command pattern; RenameOperation => RenameTracker and a subclass of this
* DelayedUpdateRenameTracker will update metastore with destfiles on  a falure, while leaving source alone.
* now creating a directory tree for each rename test, so that we can verify that it works for >1 entry deep. Size of tree depends on -Dscale flag.

The delete cleanup tries to work out from the metastore if a parent dir is empty or not; needs some isolation for its own test.

New tests all happy.

This still only updates the store on the completion of a successful rename, so the window of inconsistency (and failure ) is O(files* data); moving to do it after every file is copied
will (a) reduce peak load and (b) shrink the window significantly

Change-Id: I33c82856107f5c9c685c72eedd5309d6489a56dd