HADOOP-15183: tracking ancestors across move operations.
This adds the notion of an operation state which can be preserved across move() calls and so allow the metastore to update its view of the world without having to talk to any remote service.

For DDB it maintains that hash table of ancestors and so a set of move operations spread across parallel threads will still share the same map of ancestors to avoid creating duplicates of.

now, the bad news, as observed during test runs with debug logging: way too many metadata entries are being created. It seems to me that every putItem() call creates an ancestor list which then puts up all the ancestor markers, without checks for them existing. That is: the deeper you write, the more (expensive) write IOPs you create. If I haven't misunderstood it (and I hope I have!) then that map of ancestors isn't that useful, as entries get created anyway. All that's happening is that the write amplifcation isn't quite so big.

Added instrumentation of low level record read/write requests for dynamodb.

I'm still thinking about how best to deal with this, especially across a series of write operations likely to write to the same directory tree (that's rename and commit).

Probably something off
* track which parent paths we know to exist
* and which have been created
This would be part of the current ancestor state and, inevitably, have to be used during batch commit jobs too, so we'll know to avoid so much.

For write and commit, it'd be best if finishedWrite() did the walk up the tree and stop on the first successful probe for a parent, as its isolated, and DDB GET costs less than PUT.
That'd avoid the big write-context-spanning-operations change I'm considering but don't want to add in this patch, which is big enough anyway

Change-Id: Idabb6100e90629ddb3e04dbc3e66573abaedba13