HADOOP-15183: bulk metastore operation performance/scale
convert the bulk operation for move into into an optional BulkOperationState for put operations in the metastore.

This is then used all the way through to commit operations so as to address HADOOP-15604.

I've not yet completed up all the wiring in the metastore: This is the first step, having an optional bulk update

Change-Id: I0275a344715eab002024f1644db8858059c71bb8
Also: revert changes in ContractTestUtils as it created link conflict, and because that library gets used in contract tests by external stores, I don't want to needlessly break things.