HADOOP-15183: making addAncestors aware of the BulkOperationState
*Tests are failing; this is an interim patch*

This makes S3Guard.addAncestors aware of the state of the bulk operation by pushing down the whole ancestor search and add problem into the store itself.

Because each store can have its own model of bulk operation state, it can use this (on a bulk operation, obviously) to skip any queries of a remote DB to see if a path exists. And when there's a fallback to a get(path), and that finds a path entry, that's added to the state; no need to check again.

- we could go better and add all parents of a found path to the bulk state; that way all ancestor creation in a put(path) will know to skip the parent entries
- which may imply changing the contents of the DDBMetastore's AncestorState

I've had to push the addAncestors() call into the metastores themselve to get this to work, as only DDB maintains a state map. Alternatively, we expose some contains(path), add(path) entry to the base class (Better: make that an interface; have the impl do a set<Path>) from the outset. Although the DDB AncestorState stores the state as a map of Path -> DDBPathMetadata, its only existence which is checked for.

Change-Id: I64bf0bc372654af0471e13c324a7af0454615960