HADOOP-16357 Terasort failing against S3A
This changes the default conflict resolution mode (in code, core-default and docs) to append, to be consistent with FileOutputCommitter. Also checks for all conflict modes that the destination doesn't exist; if not calls mkdirs. If it does exist, verify that it is a directory.

Committer test  doesn't set partition policy or conflict mode, so verifies the defaults work (the test was changed before the source tests, to verify this was true)

Most the change is actually to the mocking tests, because of the switch from FS.exists(path) to getFileStatus() + mkdirs. this is why I hate mocking so much

Change-Id: I6fe9a12eea702cf75b35b8c7fb0db3f185d3f40b