HADOOP-15183 add OperationState parameter passing down the action.
* Add HADOOP-16364: translate timeout on destroy() into new exception: TableDeleteTimeoutException.
* ITestDynamoDBMetadataStore performance. Coalescing three tests into one, and inserting a new check for version marker missing a version field. The tests
which create and destroy tables are a key part of the bottleneck here, which is where the coalescing makes a big difference. FWIW,
I think we could actually use the shared table for the version checking and just restore it after, to save 60s or so.
* Trying to recreate the failure I saw where Prune failed because of an inconsistent ancestor write.

I can't seem to replicate the failure (and I deleted the table with the problem). There's now scope in the metastore to decide what to do if there is an inconsistency (throw vs warn), and I'm wondering if we need to use the operation state to choose how to react. That is: on a duplicate, inconsistent prune, do "something" else. But what?

Note: if inconsistencies in the store can be amplified by prune, this is not a new problem. It's by tracking the ancestor state that it is being detected.

Change-Id: I2e83d986a1f072d98251443e0f9c65c7036ba8ce