HADOOP-15183: Factor out the rename code from S3A FS into its own RenameOperation class, as requested.
This does provide just one place to look at the code. There are eight methods in S3AFileSystem used during the rename. I've created an interface for this and the inner class for S3AFS which bridges to them.

Looking at the methods you can see what things we should export in a lower-down layer in the S3AFS refactoring -ideally these should all be one level down.
This takes most of the new code out of the S3AFS, though the new callback interface adds some again. What is key is that

1. The new algorithm for renaming is in its own class, with src and dest params all as final fields.
2. broken up into separate methods for file and dir rename
3. and with helper methods for queuing operations etc.

The StoreContext has backed off from lambda-expressions to invoke S3AFS ops as they were getting too many, moving to an interface and again, an implementation, ContextAccessors
This adds some more code in the S3AFS, but it makes it easier to see how the methods are being used, while still allowing tests to provide their own mock implementation class.

+ InternalConstants class for internal constants
+ Move the FunctionsRaisingIOE over to fs.impl. With the move to interfaces over l-expressions these aren't being used so much, but can be picked up elsewhere. Marked as private/unstable
* slightly better diags for the AbstractCommitTerasortITests on a missing _SUCCESS marker; and a cleanup operation at the end to delete the files (the normal per-fork paths aren't used so they can avoid deletion.
* AbstractStoreOperation implements getConf() as it's that common to use.

Change-Id: I9e4420d343fb87422779c11ce89fe7710edd180c