HADOOP-15183 ensuring parent entries are correctly set up.
* DynamoDB.completeAncestry() sets the TTL on entries it creates
* DynamoDB.addAncestors() doesn't just stop at the first entry, it goes up the tree.
  When it finds a tombstone or missing entry as a parent of a vaid entry it logs @ warn and adds to the list

As a result, there's now O(depth) gets in every finishedWrite, where before it stopped at the first entry (bad)
but in completeAncestry() a put was being done at O(depth) anyway.

S3AFileSystem.finishedWrite() calls addAncestors before put().

For a bulk commit, because a bulk operation spans the add and the put, there's no duplication, the cost of a write is less:
its O(depth) with the put operations for only those files.

For a normal single file write we can do the same by creating a temp bulk write instance and using it purely for the single operation.
It does seem overkill, but it lets us glue together both operations in the sequence, which is the whole point.

Change-Id: I48ad7b2657b0d14fffc0318934af73bde8368482