HADOOP-15183: fix intermittent test; style cleanups, finishedWrite add ancestors and put call are integrated
finishedWrite() now creates a BulkUpdate if one wasn't already present, closes it afterwards. This is to ensure that the findings of the addAncestors call are used in the putAndReturn call, which will not add a PUT request for all entries we know exists. Makes DDB cost of writing a single file depth * GET + (1+ missing parent count) * PUT. Before: depth * PUT as well as extra GET/PUT calls in addAncestors. PUTs cost more than GET calls, so this is a net saving

Failing test ITestCommitOperations was tracked down to clock skew triggering a writeback of the getFileStatus result on the probes after the first commit, so causing an intermittent failure in parallel test runs (under load == worse skew).

Filed HADOOP-16382 for the underlying issue; for now simply resetting the MetricDiff counter after the various probes.

Change-Id: I85f60bc517cb0ae683961607f1f48b6f35a7004b