HADOOP-15183: DynamoDBMetadataStore tuning.
* minimising diff between trunk and branch
* completeAncestry doesn't break on first ancestor found, it continues up the path.

This is due diligence: I haven't encountered problem which arise from not doing this, I'm just making sure that we make sure those parents exist. Because operations now span >1 write, and the normal file write includes the addAncestors check which builds up the same list including with probes for the files actually existing.

Change-Id: I2edf9de75ea2546de7f97322ee0bcf838dd7591b