HADOOP-16357 fix ITestDirectoryCommitProtocol test runs by explicitly clearing bucket options.
Somehow locally the commit confict mode fs.s3a.committer.staging.conflict-mode was retaining as fail, even with the core-default.xml set. Extra tests to log the origin of the option on an assert failure implied somehow this was coming via test/resources/core-site.xml and any transitive XMLIncludes from there.

resetting the per-bucket options was enough to make this "go away"; the test is retained in case someone else's setting is similar.

Change-Id: I3a882919a78438b22fa90e9bceb516a4de640082
Note: if the core-site actually sets it for all buckets, rather than per-bucket, then the new test will fail. Why not clear it? Because we want to verify that the default is now append.