HADOOP-16430: incremental delete of S3Guard table
This patch has a BulkOperationContext passed into the delete calls, which then gets added as another arg for various operations (S3AFileSystem.removeKeys, ...) and so ends up changing too many uses/tests. Given that currently this is only used at the log level, think this is overkill.

All we need right now is the changes to

* Metastore API and impls to take a delete(Collection<Paths>) arg.
* innerDelete to call this during the paged delete for incremental deletes when > 1000 entries arre being deleted.
* ITestDeleteManyFiles to be configurable with size of files independent of the scale.test.operation.count field.

Expect a leaner PR to follow & view this one as an initial PoC.

Change-Id: I254f79eb09708e0aa12d2ed0ea50202c60fb0416