HADOOP-16380: Tombstones
This merges gabor's PR about the root dir with my test case, reworking the test case to work with subdirectories.

Where, sadly, it shows the problem exists.

If you have a file under a tombstone and that tombstone is the first item in a directory listing, then the directory will be mistaken for being empty.

Change-Id: Id02e821539f319491bd859fe0596913ce5d38287