HADOOP-16380: explicit statistics check of LIST operation
uses MetricDiffs to verify that the getFileStatus call doesn't issue a GET or LIST to S3.

This ensures that the test success isn't due to some change in how tombstones and s3 entries are resolved -it is because, when DDB is used, there are no queries to the filesystem at this point. This is a better probe as it verifies the desired behaviour rather than just the outcome

Change-Id: I6ee719ebfc95698e867972685cfad478d7f27ec2