HADOOP-16499: tuning tests to cope with a longer retry interval
it turns out that with a long retry interval, various tests which were expecting failures, be it in network setup or OOB tests were using the default retry times -extend them and tests  take longer, sometimes so much that they time out.

This patch cranks back the retry intervals on more tests to fail much faster than before, which is actually an improvement.

Considered but not done: changing the defaults in the AbstractS3ATestBase; It seems better to use the real defaults and identify those tests which expect to retry a lot

Change-Id: I1e9bac752907a6ced8a0bc7c37c95006dd8d6e98