HADOOP-16430.  delete takes place in separate thread.
This uses an executor to run the delete calls to S3 and DDB; you can also set the page size of deletes to be smaller, which will help with testing.

Side issue: moved all internal constants into the s3a.impl.InternalConstants class, and deleted the fs.s3a. one. That one only came with the OpenFile code, so nobody has been ignoring the warnings and using it anyway. At least, not in production.

This patch does not do anything about the inconsistent S3 list problem; still thinking about the best way to do this.

Change-Id: I5e626e10b7d15cb01d42dd5691573601605348e6