HADOOP-16430 checkstyle and a test
-various checkstyles
-move the parent dir cleanup in AbstractContractGetFileStatusText into the right place
-removeKeys() to switch from returning Optional<DeleteObjectsResult> into returning
 a simple DeleteObjectsResult or null
-first test of OOB cleanup, showing that in auth and nonauth we find files

That test also finds a long-standing issue: for a simple empty dir marker we don't look for child entries, ever.

And, for bonus fun, if your parent dir doesn't exist then with S3Guard you don't get found in the first place. This means that
there limits to the oob cleanup we can do

* even for an empty dir marker, at the cost of a LIST we can find and delete children
* if s3guard says the dir isn't there, the dir isn't there and there'll be no attempt to go the underlying fs.

Change-Id: Ia57f0f990d9abaca8bc9af3b48c1b7c44721dc7b