HADOOP-16207 test stabilisation
Tracked down bug where staging dir was in a temp path under the NM's dir tree, so even though the fs was shared, only tasks running on the same NM had any output to commit.

fix uses java tmp dir in junit process to set the staging dir property; also adds assertions for size of success file count so the problem is picked up on the specific job, rather than have the follow on job fail with a -1.

Change-Id: Iab0ebdbf8c87518c4274eb2d9bf8468670b957d7
but: runs are now creating a tmp/ dir under hadoop-aws for staging files. This is wrong -we should be picking up the specific one for that test fork