HADOOP-16430 pulling out copy-and-paste commonality of RenameOperation and DeleteOperation into an abstract superclass ExecutingStoreOperation
This just contains the execute() call and an check for reentrancy, but also declares a common concept: this is an operation which is executed once. Params
are set in the constructor (maybe even a builder if someone wants that) and then execute() is called.

Its overkill for just two operations, but I have a plan to move more over this way, with innerMkdirs and a fixed innerCopyFromLocalFile the obvious next stages.

Change-Id: I751cf83f2678daf9d51ce0f3d7f912d7573ba375