HADOOP-16570 expand thread leakage checks
Based on some hints from Kevin Risden, two test suites
now verify that after the test run there are no unexpected
threads -after we strip out a set of known-but-unstoppable
threads we can expect to see.

This is used in ITestS3AClosedFS to verify that an FS
instance doesn't leak any threads -useful for future regression
testing.

For the S3A committer tests we only scan for outstanding committer
pool tests; the rest are unimportant.

Also
-some comments/diags in the instrumentation class
-s3a FS close() also explcitly ishuts down the threadpools.
With a min size of 0 they'll eventually stop anyway, this just
guarantees it happens faster.

Change-Id: I2081e327ac8fb57eb38a3d119f02efce6232bad2