HADOOP-16570. Complete (mostly) the move two parallel Reading of .pendingset files
* there are pendingset-level commit/abort/revert operations to manage committing
  work and the (best effort) rolling back from failures.

* The Tasks API is used within these operations to choreograph
  commit/abort/revert actions. However, no thread pool is currently created for
  that work. I didn't want to use the pool which schedules of the file loading,
  as deadlock would have been inevitable. A separate thread pool could be
  created.
  However, unless it was actually bigger than the current pool, there
  would be no extra parallelisation.
  One special case: there were only a few tasks but they generated many, many
  files. I'm not worrying about that.

* The mock test of the committers have been reworked for this world,
  including explicitly creating and saving multiple .pendingset files to
  better stress the commit process.

* I've also moved to AssertJ assertions while trying to debug mismatches
  between the expected and actual values.
  I'd split the blame for those failing tests equally between setting up
  the mock state and me getting revert and abort to work as the existing test
  cases expected.
* some more detail in the Java docs to explain what is going on.

Regarding the state of the patch, the tests all happy;
now want to see what Yetus says. I also want to make
PartitionedStagingCommitter.replacePartitions() do its read of all .pendingset
files in parallel, so it can build up the list of partitions to replace a
bit faster when there is the output of many tasks to process.

I'm also actually wondering what it would take to use the MockS3AClient here
across more tests. Currently it is good (and with this patch better) at
Simulating incomplete multipart uploads -including with tracking of active
uploads. We could probably expand this to model more of the final state of the
store -for example actually simulating the persisted state of the store.
Worth a thought -though it is probably a moving target.

Change-Id: Idfa8198a920664f2fefe441d317b8e0fb681d368