HADOOP-16697. S3Guard Import to add a -auth command
This will cause the meta store to marks the directory tree as
authoritative once the import has completed its scan.

It is off by default because you have to be 100% confident that
during a potentially slow scan nobody is writing to that directory tree.

This change includes moving the core operations of S3GuardTool.Import into
an Importer class for ease of testing and so it can live alongside the auditor.
They both use the same (new) MetadataStore method
markAuthoritative(path, BulkOperation).

For DDB, the AncestorState is used to determine what to mark
-we don't attempt any listing of the directory tree.

For Local S3Guard, is currently a no-op. I'm not worried about that,
Especially as the auditor only audits DDB Tables.

The documentation is yet to be updated!

Change-Id: Ifc47713d3b70a6bcf2671c7e2ded31a15ee4d9c6