HADOOP-16697 Auth mode: imports
More work on import, with a verbose mode to print out what is going on.

When importing a listing and we find a file -we now query the store and only update the entry if the file is different from before, where different == new timestamp, etag, or length.

Now, we can get timestamp differences due to clock skew in PUT operations: we can recognise this because the etags are the same. And, the recursive list operation doesn't retrieve the versionID, the existing entry may in fact be more complete. So in the special case of updating due to clock skew and Etag is the same, we copy over the version ID to the new entry.

Looking at the S3Guard operations log in these test runs has shown that some of my "clever" optimisations over the summer where in fact "utterly wrong".

Specifically, when adding an entry to the store, we PUT all the parent directory markers -without looking to see if they exist. As a result, any parent directory marked as auth is immediately marked as unauth if any file or directory is added anywhere underneath.

Will fix in this (growing) PR.

Change-Id: I315f342a1aff4d633cd9c4edb202237358e76dee