HADOOP-16697 Auth mode: scan parents before PUTting updates
This is probably slightly slower but put less write load on the store so probably
not too bad -especially as we do the ancestor state scan first; there's no
duplication going to take place within the same operation.

Fixes all failing tests to work, except now prune of non tombstone
is *not* marking the parent as non-auth.

Change-Id: If65337e08b4941712f7f0c01b09d605f97a1f3a1