HADOOP-16697 auth mode and empty dirs; pruning tuning
- empty dirs are to be marked as auth when creating with mkdir or
when adding a fake dir. This is still failing its tests.

- prune improvements. There's an issue which has surface there,
namely: we must not prune directory entries because they may have children
and if we prune them then those entries become orphan.

The only way to prune dirs would be for each dir to ask for 1+ child and
if none were found: delete. This is out of scope for this PR.

Change-Id: I6a0bc6ec3e0046093eed072bb44f28db9b81ff5d