HADOOP-16729. Extract version numbers to head of pom.xml
For branch-3.2

To be able to easily replace third-party dependency version numbers it
would be useful to collect their version numbers at the top of the pom.xml.

For many third-parties (e.g. slf4j, jetty, etc.) this is already done,
but I would need the same for others. The change doesn't have any effect
on the code or the build, no version numbers would be changed.