HADOOP-16697 preparing for review.
* tests for the CLI, command line parsing etc
* tuning CLI based on manual use (mainly logging)

S3GuardTool is now closeable, and will close its FS and MS in close. Without this, our tests can leak FS and MS instances. We just hadn't noticed.

The CLI entry point of S3GuardTool does the right thing and closes the tool; for our tests we have to explicitly do it. This is done in a mixture of try-with-resources and building up a list of tools to close in test tear down. All tests which patch in an existing metastore need to replace it with a null one before the close operation. Are you

Change-Id: I061eb0ece24e84aaffb4f12a3e7c920011146dd6
Testing: new tests are good, doing more regression testing. Can't run -Dlocal for some reason.