HADOOP-16697. better logging of import
Based on Gabor's feedback
-rename a method
-review import logging and extend

the -verbose mode already logs each file/dir; now non-verbose and
verbose both print summaries of the operation: the number of files
and dirs added.

Note, well looking at the logging at debug level I can see that after the import, when we mark everything as authoritative, we end up overwriting all the dir entries just added.

It's not that easy to avoid doing this while keeping the intermediate state of the S3Guard tables consistent; I'd have to modify that marking operation to do a Get on each dir and only update non-auth ones, while tagging the dirs as auth when created.

It's doable; we'd be moving from two writes to one read, one write, or somehow blue together the tracking even more.

I'm not convinced it's worth the extra complexity for an operation which is not expected to be run very often.

Change-Id: Ic2c6378d4c47fe2dc36d8f6d6b7d5e50735ceacb