HADOOP-16875: AbfsOutputStream.close() wraps IOEs with a new instance of the same class.
This moves the wrapping logic out of flushInternal into close(),
using IOUtils.wrapException to ensure the type of the wrapped exception
is the same as that caught (without that the E2E exceptions fail).

the test uses try-with-resources as suggested by Thomas; we delete a file while
it is being written; this causes hsync to fail -with close() also throwing
an exception which is then caught and suppressed.

Change-Id: If1bcd2885bccbf39de888b8295a5a7e529935f2d