HADOOP-16746. Authoritative Directories
This fixes the problem wherein a listStatus of a parent dir
would stamp on the isAuthoritative flag of every child entry.

This was caused by us blindly overriding entries with new ones.

This patch

* avoids writing unchanged entries by building a list of those and passing it
  down to the meta store in the put(DirListMeta) call.
* in DDB: filter out of those entries.
* in non-auth mode, build up a list of entries to add, and write in a batch
  at the end. This is more efficient than the one-by-one operation which
  was being performed, especially as there is no BulkOperationState to cache
  previous work.
* adds tests to verify no DDB writes take place on repeated lists
* fixes up all calls of put() to handle the new list of unchanged entries
* Fixes a bug in TestS3Guard which caused NPEs

The code in S3Guard.dirListingUnion() is a bit ugly as it
has two very different sequences (auth vs nonauth) intermingled.
We should consider splitting up the two union mechanisms,
so as to make it easier to understand how the different update/unions
work.

Change-Id: If5f2427f49a46cb7827f1634b6528ddd0e265778